Add V1 Latex
This adds latex that generates the V1 whitepaper. The original latex was lost and reconstructed.

This commit also includes an updated build_latex.yml with a more recent latex compiler. This reconstructed latex generates an almost identical PDF. Notably, every page contains the exact same words and every line ends with the same character. The only differences exist due to unavoidable compiler changes and are as follows:
Margins are increased slightly.
Spacing between lines is reduced slightly to compensate.
A couple footnotes that used to overflow into the right column are now fully contained in the left column.